```text feat: Add loading message during form submission
The code changes in this commit add a loading message during form submission. When the form is submitted, the loading message is displayed using the 'loadertext' class. This provides visual feedback to the user that the form is being processed.

The changes include:
- Adding the 'loadertext' id to the loading message element in the 'test.html' file.
- Setting the display property of the loading message element to 'block' when the form is submitted in the 'script.js' file.
- Setting the display property of the loading message element to 'none' when the form submission is complete in the 'script.js' file.

This commit addresses the need to provide a better user experience by indicating that the form submission is in progress.